+++ lib/php/src/PaginationHelper.php
@@ -30,7 +30,10 @@ class PaginationHelper
         $end_page = $end_page > $total_page ? $total_page : $end_page;
 
         $link = parse_url($request->server->get('REQUEST_URI'), PHP_URL_PATH);
-        $query_string = UrlHelper::buildQuery(array_filter($request->query->all()), ['page' => null]);
+        $queries = array_filter($request->query->all(), function ($query) {
+            return !empty($query) || $query == '0';
+        });
+        $query_string = UrlHelper::buildQuery($queries, ['page' => null]);
         if (empty($query_string)) {
             $link .= '?';
         } else {